updated for Apple Silicon
* Updated for macOS Ventura and Xcode 14.2
* Apple Silicon compatible
* updated project to macOS 11+
* fixed most warnings and deprecations
* replaced ShortcutRecorder Framework with MASShortcut package
* general refactoring
* updated Preferences window layout
* replaced most static images with SF Symbols
* moved remaining images to Assets catalogue
* updated icon to user submitted icon https://github.com/2ndalpha/gasmask/issues/203
* refreshed status menu icon
* refreshed .hst icon
* added support for restarting mDNSResponder via NSUserAppleScriptTask
* updated .gitignore to ignore .DS_Store files
* updated README to indicate support for macOS Big Sur and above

+++ Source/CombinedHostsPredicateController.h
@@ -26,7 +26,7 @@
     IBOutlet NSScrollView *lowerScrollView;
     IBOutlet NSView *parentView;
     @private
-    int rowCount;
+    NSUInteger rowCount;
     CombinedHosts *selectedHostsFile;
     NSImageView *hintView;
 }
